refactor: Custom View들 파일 분리 및 PreviousViewModel 생성
